Extract 3 library-operation handlers to importer_operations
Sprint 4, PR #7.  Second bite of the prepare_import_payload monster.
Moves the trio of nested-closure operation handlers (mass update,
toggle-select, delete-collections) out to their own module.

## What moved

Three nested closures inside prepare_import_payload -> module-scope
functions in modules/importer_operations.py:

- handle_mass_update_operation        (73 lines)
- handle_toggle_select_operation      (85 lines)
- handle_delete_collections_operation (79 lines)

Total: 237 lines lifted out of the mega-function.

## Handler signatures

Each was previously reaching into enclosing scope for:
- libraries_data     -- per-library flat dict being populated
- report             -- ImportReport collecting mapped/unmapped paths
- mass_update_defs / toggle_select_defs -- op-type dispatch tables

Those are now explicit keyword-only arguments.  Positional args are
still the four per-call values (lib_id, lib_name, op_key, op_value),
keeping call sites readable.

Leading '_' stripped from the extracted names because the handlers
are now the public API of importer_operations (still module-internal,
just not per-function-internal).  Return convention unchanged:
(handled, imported) tuple where handled=True means 'this handler
owned the op_key, stop dispatching' and imported=True means at
least one value was mapped into libraries_data.

## Tiny helpers moved with them

_encode_json, _clean_custom_value, _normalize_op_items were also
pulled into importer_operations.  They're only called from the
operation handlers, so DRY says single source of truth beats
'accessible from importer.py' -- the module-scope copies in
importer.py (hoisted in PR #1507) are DELETED as part of this PR.

## Import strategy

modules/importer.py now imports importer_operations as a MODULE
(not name-by-name).  Call sites use importer_operations.handle_X(...)
which makes the operation-handler cluster visually obvious and
keeps the three fat kwargs from cluttering the import list.

modules/importer_operations.py imports ImportReport only for the
TYPE_CHECKING type-hint -- no runtime import of modules.importer,
so zero risk of circular imports.

## No behavior changes

Byte-identical mapping of every op_key + op_value combination.
No test changes required -- the handlers were nested closures
that could only be reached through prepare_import_payload, so
no test targeted them directly and none needed updating.

## Impact

- modules/importer.py: 1363 -> 1131 (-232 / -17.0%)
- modules/importer_operations.py: NEW 333 lines
- prepare_import_payload body: 958 -> 748 (-210 / -21.9%)

Cumulative importer.py progress across Sprint 4:
  2027 -> 1131 = -896 / -44.2% (roughly half the original file)

All 1285 tests pass.  Ruff + black clean.
